WIP: getTotalKilometrage moved to service layer;
